#12 adjust tests to clean architecture
